+++ core/src/test/java/org/imec/ivlab/core/vaccination/VaccinationMappingReaderTest.java
@@ -1,13 +1,15 @@
 package org.imec.ivlab.core.vaccination;
 
 import java.util.List;
-import org.junit.jupiter.api.TestCase;
 import lombok.extern.log4j.Log4j2;
 import org.apache.commons.lang3.StringUtils;
 
-@Log4j2
-public class VaccinationMappingReaderTest extends TestCase {
+import org.testng.annotations.Test;
 
+@Log4j2
+public class VaccinationMappingReaderTest {
+
+  @Test
   public void testReadMappings() {
 
     List<VaccinationMapping> vaccinationMappings = VaccinationMappingReader.getInstance().getVaccinationMappings();
