+++ super/pom.xml
@@ -51,12 +51,6 @@
             <artifactId>log4j-core</artifactId>
         </dependency>
 
-		<dependency>
-			<groupId>org.junit.jupiter</groupId>
-			<artifactId>junit-jupiter</artifactId>
-			<scope>test</scope>
-		</dependency>
-
         <dependency>
             <groupId>org.hamcrest</groupId>
             <artifactId>hamcrest-core</artifactId>
@@ -115,14 +109,6 @@
                 <version>7.5.1</version>
                 <scope>test</scope>
             </dependency>
-
-			<dependency>
-				<groupId>org.junit</groupId>
-				<artifactId>junit-bom</artifactId>
-				<version>5.10.0</version>
-				<type>pom</type>
-				<scope>import</scope>
-			</dependency>
 
             <dependency>
                 <groupId>org.apache.logging.log4j</groupId>
